# ENB: not build templates for JavaScript