Resolved #1134 

Assembly became a little faster.

               | before    | after
------------ | ----------- | -----------
bundles   | 2361ms  | **2140ms**
examples | 6778ms | **5573ms**